Clin-413 fix icon sexe fetus
